removed renderFilmCard func to movie-controller, add renderFilmCards
